requirements.txt
# Core app dependencies
flask
werkzeug
pillow
numpy
pandas

# ML dependencies
torch
torchvision

# Reporting
reportlab

# Optional (deployment)
gunicorn
